requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
-c constraints.txt
accelerate>=0.25.0
build
colored
cuda-python # Do not override the custom version of cuda-python installed in the NGC PyTorch image.
diffusers>=0.27.0
lark
mpi4py
numpy<2
onnx>=1.12.0
onnx_graphsurgeon>=0.5.2
openai
polygraphy
psutil
nvidia-ml-py>=12,<13
# Just a wrapper since nvidia-modelopt requires pynvml
pynvml==12.0.0
pulp
pandas
h5py==3.12.1
StrEnum
sentencepiece>=0.1.99
tensorrt~=10.13.0.35
# https://docs.nvidia.com/deeplearning/frameworks/pytorch-release-notes/rel-25-05.html#rel-25-05 uses 2.8.0a0.
torch>=2.7.1,<=2.8.0a0
torchvision
nvidia-modelopt[torch]~=0.31.0
nvidia-nccl-cu12
nvidia-cuda-nvrtc-cu12
transformers==4.53.1
pydantic>=2.9.1
pydantic-settings
pillow==10.3.0
wheel<=0.45.1
optimum
# evaluate needs datasets>=2.0.0 which triggers datasets>3.1.0 which is not stable: https://github.com/huggingface/datasets/issues/7467
datasets==3.1.0
evaluate
mpmath>=1.3.0
click
click_option_group
aenum
pyzmq
fastapi==0.115.4
uvicorn
setuptools<80
ordered-set
peft
einops
flashinfer-python==0.2.5
opencv-python-headless
xgrammar==0.1.19
backoff
nvtx
matplotlib # FIXME: this is added to make nvtx happy
meson
ninja
etcd3
blake3
llguidance==0.7.29
